Refactor package name from `builders` to `builder`
