+++ play-services-constellation/core/src/main/kotlin/org/microg/gms/constellation/core/proto/builder/ClientInfoBuilder.kt
@@ -1,6 +1,6 @@
 @file:RequiresApi(Build.VERSION_CODES.O)
 
-package org.microg.gms.constellation.core.proto.builders
+package org.microg.gms.constellation.core.proto.builder
 
 import android.annotation.SuppressLint
 import android.content.Context
