+++ play-services-constellation/core/src/main/kotlin/org/microg/gms/constellation/core/proto/builder/RequestBuildContext.kt
@@ -1,4 +1,4 @@
-package org.microg.gms.constellation.core.proto.builders
+package org.microg.gms.constellation.core.proto.builder
 
 import android.content.Context
 import org.microg.gms.constellation.core.AuthManager
